src/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "avro-python3>=1.10.2",
    "kafka-python>=2.2.7",
    "markupsafe<3.0.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "opencv-python>=4.8.0",
    "httpx>=0.28.1",
    "scipy>=1.15.3",
    "torch>=2.7.1",
    "lap>=0.5.12",
]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"


[dependency-groups]
dev = [
    "deptry>=0.23.0",
    "ruff>=0.11.5",
]
